fix: move pool utils from `types.ts` to separate file
